Adjustments to debruijn_graph to support bgl adapter
node_rank is understood with this commit to have 1’s in the FIRST edge
of each node in the table (as opposed to L in the paper that marked the
LAST edge of each node with a 1). Right now the flagging of edge
symbols matches W in the paper but I found minor adjustments were
needed in _forward and outgoing accordingly. I am going to investigate
whether a different approach to flagging W will resolve this.